fix(web): fix svg background in steps
